Don't install stuff that's already installed.

+++ vm/setup-python.sh
@@ -1,25 +1,31 @@
 apt-get install python-pip -y
-wget https://bootstrap.pypa.io/ez_setup.py -O - | sudo python
 
 ## Install scipy, numpy, etc.
 apt-get build-dep python-scipy -y
 apt-get install python-scipy libagg-dev python-matplotlib -y
 
-## SLICOT
-git clone git://github.com/avventi/Slycot.git
-cd Slycot
-python setup.py install
-cd ..
-rm -r Slycot
+if [ $(python -c "import setuptools" &> /dev/null) ]; then
+	wget https://bootstrap.pypa.io/ez_setup.py -O - | sudo python
+	rm ./setuptools-3.5.1.zip
+fi
 
-## python-control
-wget http://sourceforge.net/projects/python-control/files/control-0.6d.tar.gz/download -O control-0.6d.tar.gz
-tar xzf control-0.6d.tar.gz
-rm control-0.6d.tar.gz
-cd control-0.6d
-python setup.py install
-cd ..
-rm -r control-0.6d
+if [ $(python -c "import control.matlab" &> /dev/null) ]; then
+	## Install SLICOT wrapper
+	git clone git://github.com/avventi/Slycot.git
+	cd Slycot
+	python setup.py install
+	cd ..
+	rm -r Slycot
+
+	## Install python-control library
+	wget http://sourceforge.net/projects/python-control/files/control-0.6d.tar.gz/download -O control-0.6d.tar.gz
+	tar xzf control-0.6d.tar.gz
+	rm control-0.6d.tar.gz
+	cd control-0.6d
+	python setup.py install
+	cd ..
+	rm -r control-0.6d
+fi
 
 # cvxpy
 pip install cvxpy
